requirements.txt
pandas
matplotlib
pillow
opencv-python
dvc